requirements.txt
aiogram>=3.20.0
langchain>=0.3.25
langchain-core>=0.3.67  
langchain-community>=0.3.24
langchain-text-splitters>=0.3.8
langchain-chroma>=0.2.4
langchain-gigachat>=0.3.10
langgraph>=0.6.5
python-dotenv==1.1.0